feat: 기존 controller test 통합 테스트 / 단위 테스트 분리
