+++ src/test/java/roomescape/support/ControllerTestSupport.java
@@ -1,9 +1,10 @@
-package roomescape.auth;
+package roomescape.support;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.jdbc.Sql;
@@ -11,15 +12,18 @@ import roomescape.config.TestTimeConfig;
 
 import java.util.Map;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@Sql(scripts = {"/truncate.sql", "/test-data.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
-@Import(TestTimeConfig.class)
 @ActiveProfiles("test")
-abstract class AuthApiTestSupport {
+@Import(TestTimeConfig.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@Sql(scripts = {"/truncate.sql", "/test-data.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
+public abstract class ControllerTestSupport {
+
+    @LocalServerPort
+    private int port;
 
     @BeforeEach
-    void setUp() {
-        RestAssured.port = 8080;
+    void setUpRestAssuredPort() {
+        RestAssured.port = port;
     }
 
     protected String loginToken(String name, String password) {
@@ -35,6 +39,18 @@ abstract class AuthApiTestSupport {
 
     protected String loginUserToken() {
         return loginToken("a", "test1");
+    }
+
+    protected String loginWaitingUserToken() {
+        return loginToken("b", "test2");
+    }
+
+    protected String loginOtherUserToken() {
+        return loginToken("c", "test3");
+    }
+
+    protected String loginManagerToken() {
+        return loginToken("d", "test4");
     }
 
     protected String bearer(String token) {
